+++ packages/parameters/src/ssm/getParameter.ts
@@ -1,0 +1,17 @@
+import { SSMProvider, DEFAULT_PROVIDERS } from './SSMProvider';
+import type { SSMGetOptionsInterface } from 'types/SSMProvider';
+
+const getParameter = (
+  name: string,
+  options?: SSMGetOptionsInterface
+): Promise<undefined | string | Record<string, unknown>> => {
+  if (!DEFAULT_PROVIDERS.hasOwnProperty('ssm')) {
+    DEFAULT_PROVIDERS.ssm = new SSMProvider();
+  }
+
+  return DEFAULT_PROVIDERS.ssm.get(name, options);
+};
+
+export {
+  getParameter,
+};
