+++ packages/parameters/src/ssm/index.ts
@@ -1,0 +1,4 @@
+export * from './SSMProvider';
+export * from './getParameter';
+export * from './getParameters';
+export * from './getParametersByName';
